requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
autopep8==1.5.7
billiard==4.2.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
daphne==4.1.2
dj-database-url==0.5.0
Django==5.1.6
djangorestframework==3.16.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
kiwisolver==1.4.8
kombu==5.5.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
prompt_toolkit==3.0.50
psutil==7.0.0
py-cpuinfo==9.0.0
py-srt==1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.7.0
pycparser==2.22
pyOpenSSL==25.0.0
pyparsing==3.2.3
pysrt==1.1.2
python-dateutil==2.9.0.post0
python-decouple==3.4
pytz==2021.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
service-identity==24.2.0
setuptools==76.0.0
six==1.17.0
sqlparse==0.4.1
srt==3.5.3
sympy==1.13.1
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0+rocm6.2.4
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.134
ultralytics-thop==2.0.14
Unipath==1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==5.3.0
zope.interface==7.2
